+++ packages/lib/server/service/__tests__/InsightsRoutingService.integration-test.ts
@@ -7,7 +7,7 @@ import { ColumnFilterType } from "@calcom/features/data-table/lib/types";
 import prisma from "@calcom/prisma";
 import { BookingStatus, MembershipRole } from "@calcom/prisma/enums";
 
-import { InsightsRoutingBaseService as InsightsRoutingService } from "../../service/insightsRoutingBase";
+import { InsightsRoutingBaseService as InsightsRoutingService } from "../../service/InsightsRoutingBaseService";
 
 // SQL condition constants for testing
 const NOTHING_CONDITION = Prisma.sql`1=0`;
